Authenticate to PrivacyIDEA lazily
Currently each time the plugin is executed it will try to authenticate to the
PrivacyIDEA server.  This happens because the plugin constructor instantiates
the PrivacyIdeaUtils class which calls its authenticate() method.  This is not
necessary, e.g. when embedded in the role manager.  Instead ensure
authenticate() is called before any operation on the PrivacyIDEA server.