class_mfaAccount.inc: Handle apply action if not child of management
The apply action ($POST["edit_apply"]) is handled automatically if the plugin
is a child of an instance of a management class and submitted settings are
saved.  For other cases, e.g. if the plugin is a child of an instance of the
tabs class, this needs to be handled manually by calling the save_object
method.