class_PrivacyIdeaUtils.inc: Use <config.class>->get_cfg_value_ext() instead of <config.class>->get_cfg_value().
Where as get_cfg_value() tries to use the GOsa² property system,
get_cfg_value_ext() directly reads from the gosa.conf config file.

This avoids log messages of the kind:

| debug | warning | Unconfigured property: 'core::pi<SomeThing>'

As a side note, this is wanted behaviour. The privacyIDEA service account
settings (et al.) shall not be stored in LDAP. They shall always be
obtained from the gosa.conf file, so it requires someone with filesystem
access to tinker with them.